requirements.txt
    MarkupSafe==2.0.1
    Jinja2==2.11.3
    Flask==1.1.4

    dnspython==2.2.1
    paho-mqtt==1.6.1

    typing_extensions
    eventlet
    greenlet
    Flask-SocketIO==4.3.2
    Flask-Cors==3.0.10
    urllib3==1.26.18
    requests==2.28.1
    dotenv
    requests
    werkzeug
 